Fix SHELLEXECUTEINFO object and hProcess handle leak on Windows
The `SHELLEXECUTEINFO` object was previously allocated on the heap and
never freed. We now use the stack instead to fix this memory leak.

The `SHELLEXECUTEINFO.hProcess` handle is supposed to be closed manually
when the `SEE_MASK_NOCLOSEPROCESS` flag is used. We now close the
handle after we're done with it.